swap: rename automata to process_msgs
